requirements.txt
flask
python-dotenv
openai>=1.0.0
